backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
langchain
chromadb
sentence-transformers
transformers
faiss-cpu
accelerate
python-dotenv
PyPDF2
requests
torch
python-docx
python-pptx
pillow
pytesseract
pdf2image
pandas
langchain-community